requirements.txt
streamlit
torch
Pillow
tqdm
ftfy
regex
git+https://github.com/openai/CLIP.git
